Fix to work with latest Snapchat 3.0.5 (2013-08-21)
A lot of internal stuff has changed in Snapchat in the latest updates.
Now, the getImageBitmap() method doesn't provide a context anymore and the operations depending on a context have to be done in the showImage() method, similarly to the way video saving is done. Because DRY, I streamlined the code and put the joint tasks for displaying a toast and calling the media scanner in a separate method. (Well, media scanning was in there already, that's why the method's name changed from runMediaScan to runMediaScanAndToast.)