Implemented full settings funcitonality
Added the dialog that asks whether to save the media. In case saving mode is set to SAVE_AUTO, the image will still always be saved upon opening, but if the user chooses "Dismiss" at the dialog, we silently delete the file. Only if the user chooses to save, or enter the settings, do we actually show the toast (if enabled) that the media was saved and call the media scanner.